Update config to ESLint Flat Config
