--- index.d.cts
@@ -1,4 +1,0 @@
-import { type TSESLint } from '@typescript-eslint/utils';
-
-declare const config: TSESLint.Linter.Config;
-export = config;
