+++ index.d.ts
@@ -1,0 +1,4 @@
+import { type TSESLint } from '@typescript-eslint/utils';
+
+declare const config: TSESLint.Linter.FlatConfig[];
+export default config;
