+++ package.json
@@ -20,13 +20,13 @@
     "Karl Horky <karl.horky@gmail.com>"
   ],
   "type": "module",
-  "main": "index.cjs",
+  "main": "index.js",
   "bin": {
     "upleveled-eslint-install": "bin/install.js"
   },
   "files": [
-    "index.cjs",
-    "index.d.cts",
+    "index.js",
+    "index.d.ts",
     "bin",
     "templates",
     "tsconfig.base.json"
@@ -35,6 +35,7 @@
     "postpublish": "cd __tests__ && pnpm update-config-version"
   },
   "devDependencies": {
+    "@types/eslint": "8.44.7",
     "@types/prettier": "2.7.3",
     "@typescript-eslint/utils": "6.9.1",
     "eslint-config-upleveled": "5.1.3",
@@ -64,6 +65,7 @@
     "eslint-plugin-testing-library": "^6.1.0",
     "eslint-plugin-unicorn": "^49.0.0",
     "eslint-plugin-upleveled": "^2.1.9",
+    "globals": "^13.23.0",
     "typescript": "^5.2.2"
   },
   "engines": {
