+++ types/@next__eslint-plugin-next.d.ts
@@ -1,0 +1,6 @@
+declare module '@next/eslint-plugin-next' {
+  import type { ESLint } from 'eslint';
+
+  let plugin: ESLint.Plugin;
+  export default plugin;
+}
