+++ types/eslint-plugin-jsx-a11y.d.ts
@@ -1,0 +1,6 @@
+declare module 'eslint-plugin-jsx-a11y' {
+  import type { ESLint } from 'eslint';
+
+  let plugin: ESLint.Plugin;
+  export default plugin;
+}
