+++ types/eslint-plugin-jsx-expressions.d.ts
@@ -1,0 +1,6 @@
+declare module 'eslint-plugin-jsx-expressions' {
+  import type { ESLint } from 'eslint';
+
+  let plugin: ESLint.Plugin;
+  export default plugin;
+}
