+++ types/eslint-plugin-react-hooks.d.ts
@@ -1,0 +1,6 @@
+declare module 'eslint-plugin-react-hooks' {
+  import type { ESLint } from 'eslint';
+
+  let plugin: ESLint.Plugin;
+  export default plugin;
+}
