+++ types/eslint-plugin-react.d.ts
@@ -1,0 +1,6 @@
+declare module 'eslint-plugin-react' {
+  import type { ESLint } from 'eslint';
+
+  let plugin: ESLint.Plugin;
+  export default plugin;
+}
