+++ types/eslint-plugin-security.d.ts
@@ -1,0 +1,6 @@
+declare module 'eslint-plugin-security' {
+  import type { ESLint } from 'eslint';
+
+  let plugin: ESLint.Plugin;
+  export default plugin;
+}
