+++ types/eslint-plugin-upleveled.d.ts
@@ -1,0 +1,6 @@
+declare module 'eslint-plugin-upleveled' {
+  import type { ESLint } from 'eslint';
+
+  let plugin: ESLint.Plugin;
+  export default plugin;
+}
